Started to look into splitting out the commands into separate files to make things a little easier to navigate around and maintain.
