+++ src/amnuts230.c
@@ -6208,210 +6208,7 @@ look(UR_OBJECT user)
 }
 
 
-/*
- * Show who is on.  type 0=who, 1=fwho, 2=people
- */
-void
-who(UR_OBJECT user, int type)
-{
-  char line[RECAP_NAME_LEN + USER_DESC_LEN * 3];
-  char rname[ROOM_NAME_LEN + PERSONAL_ROOMNAME_LEN + 1];
-  char idlestr[20];
-  char sockstr[3];
-  const char *portstr;
-  UR_OBJECT u;
-  int mins;
-  int idle;
-  int linecnt;
-  int rnamecnt;
-  int total;
-  int invis;
-  int logins;
-#ifdef NETLINKS
-  int remote = 0;
-#endif
 
-  total = invis = logins = 0;
-
-  if (type == 2 && !strcmp(word[1], "key")) {
-    write_user(user,
-               "\n+----------------------------------------------------------------------------+\n");
-    write_user(user,
-               "| ~OL~FCUser login stages are as follows~RS                                           |\n");
-    write_user(user,
-               "+----------------------------------------------------------------------------+\n");
-    vwrite_user(user,
-                "| ~FY~OLStage %d :~RS The user has logged onto the port and is entering their name     |\n",
-                LOGIN_NAME);
-    vwrite_user(user,
-                "| ~FY~OLStage %d :~RS The user is entering their password for the first time           |\n",
-                LOGIN_PASSWD);
-    vwrite_user(user,
-                "| ~FY~OLStage %d :~RS The user is new and has been asked to confirm their password     |\n",
-                LOGIN_CONFIRM);
-    vwrite_user(user,
-                "| ~FY~OLStage %d :~RS The user has entered the pre-login information prompt            |\n",
-                LOGIN_PROMPT);
-    write_user(user,
-               "+----------------------------------------------------------------------------+\n\n");
-    return;
-  }
-  write_user(user,
-             "\n+----------------------------------------------------------------------------+\n");
-  write_user(user,
-             align_string(1, 78, 0, NULL, "%sCurrent users %s",
-                          (user->login) ? "" : "~FG", long_date(1)));
-  switch (type) {
-  case 0:
-    vwrite_user(user,
-                "%s  Name                                           : Room            : Tm/Id\n",
-                (user->login) ? "" : "~FC");
-    break;
-  case 1:
-    write_user(user,
-               "~FCFriend                                           : Room            : Tm/Id\n");
-    break;
-  case 2:
-    write_user(user,
-               "~FCName            : Level Line Ignall Visi Idle Mins  Port  Site/Service\n");
-    break;
-  }
-  write_user(user,
-             "+----------------------------------------------------------------------------+\n\n");
-  for (u = user_first; u; u = u->next) {
-    if (u->type == CLONE_TYPE) {
-      continue;
-    }
-    mins = (int) (time(0) - u->last_login) / 60;
-    idle = (int) (time(0) - u->last_input) / 60;
-#ifdef NETLINKS
-    if (u->type == REMOTE_TYPE) {
-      portstr = "   @";
-    } else
-#endif
-    {
-#ifdef WIZPORT
-      if (u->wizport) {
-        portstr = " WIZ";
-      } else
-#endif
-      {
-        portstr = "MAIN";
-      }
-    }
-    if (u->login) {
-      if (type != 2) {
-        continue;
-      }
-      vwrite_user(user,
-                  "~FY[Login stage %d]~RS :  -      %2d   -    -  %4d    -  %s  %s:%s\n",
-                  u->login, u->socket, idle, portstr, u->site, u->site_port);
-      ++logins;
-      continue;
-    }
-    if (type == 1 && !user_is_friend(user, u)) {
-      continue;
-    }
-    ++total;
-#ifdef NETLINKS
-    if (u->type == REMOTE_TYPE) {
-      ++remote;
-    }
-#endif
-    if (!u->vis) {
-      ++invis;
-      if (u->level > user->level && !(user->level >= ARCH)) {
-        continue;
-      }
-    }
-    if (type == 2) {
-      if (u->afk) {
-        strcpy(idlestr, " ~FRAFK~RS");
-      } else if (u->malloc_start) {
-        strcpy(idlestr, "~FCEDIT~RS");
-      } else {
-        sprintf(idlestr, "%4d", idle);
-      }
-#ifdef NETLINKS
-      if (u->type == REMOTE_TYPE) {
-        strcpy(sockstr, " @");
-      } else
-#endif
-        sprintf(sockstr, "%2d", u->socket);
-      linecnt = 15 + teslen(u->recap, 15);
-      vwrite_user(user,
-                  "%-*.*s~RS : %-5.5s   %s %-6s %-4s %s %4d  %-4s  %s\n",
-                  linecnt, linecnt, u->recap, user_level[u->level].name,
-                  sockstr, noyes[u->ignall], noyes[u->vis], idlestr, mins,
-                  portstr, u->site);
-      continue;
-    }
-    sprintf(line, "  %s~RS %s", u->recap, u->desc);
-    if (!u->vis) {
-      *line = '*';
-    }
-#ifdef NETLINKS
-    if (u->type == REMOTE_TYPE) {
-      line[1] = '@';
-    }
-    if (!u->room) {
-      sprintf(rname, "@%s", u->netlink->service);
-    } else {
-#endif
-      strcpy(rname,
-             is_personal_room(u->room) ? u->room->show_name : u->room->name);
-#ifdef NETLINKS
-    }
-#endif
-
-    /* Count number of colour coms to be taken account of when formatting */
-    if (u->afk) {
-      strcpy(idlestr, "~FRAFK~RS");
-    } else if (u->malloc_start) {
-      strcpy(idlestr, "~FCEDIT~RS");
-    } else if (idle >= 30) {
-      strcpy(idlestr, "~FYIDLE~RS");
-    } else {
-      sprintf(idlestr, "%d/%d", mins, idle);
-    }
-    linecnt = 45 + teslen(line, 45);
-    rnamecnt = 15 + teslen(rname, 15);
-    vwrite_user(user, "%-*.*s~RS  %1.1s : %-*.*s~RS : %s\n", linecnt, linecnt,
-                line, user_level[u->level].alias, rnamecnt, rnamecnt, rname,
-                idlestr);
-  }
-  write_user(user,
-             "\n+----------------------------------------------------------------------------+\n");
-  switch (type) {
-  case 0:
-  case 2:
-    if (type == 2) {
-      sprintf(text, "and ~OL~FG%d~RS login%s ", logins, PLTEXT_S(logins));
-    }
-#ifdef NETLINKS
-    write_user(user,
-               align_string(1, 78, 0, NULL,
-                            "Total of ~OL~FG%d~RS user%s %s: ~OL~FC%d~RS visible, ~OL~FC%d~RS invisible, ~OL~FC%d~RS remote",
-                            total, PLTEXT_S(total), type == 2 ? text : "",
-                            amsys->num_of_users - invis, invis, remote));
-#else
-    write_user(user,
-               align_string(1, 78, 0, NULL,
-                            "Total of ~OL~FG%d~RS user%s %s: ~OL~FC%d~RS visible, ~OL~FC%d~RS invisible",
-                            total, PLTEXT_S(total), type == 2 ? text : "",
-                            amsys->num_of_users - invis, invis));
-#endif
-    break;
-  case 1:
-    write_user(user,
-               align_string(1, 78, 0, NULL,
-                            "Total of ~OL~FG%d~RS friend%s : ~OL~FC%d~RS visible, ~OL~FC%d~RS invisible",
-                            total, PLTEXT_S(total), total - invis, invis));
-    break;
-  }
-  write_user(user,
-             "+----------------------------------------------------------------------------+\n");
-}
 
 
 /*
@@ -6545,134 +6342,7 @@ display_files(UR_OBJECT user, int admins)
 }
 
 
-/*
- * Show the list of commands, credits, and display the helpfiles for the given command
- */
-void
-help(UR_OBJECT user)
-{
-  char filename[80];
-  const struct cmd_entry *com, *c;
-  size_t len;
-  int found;
 
-  if (word_count < 2 || !strcasecmp(word[1], "commands")) {
-    switch (user->cmd_type) {
-    case 0:
-      help_commands_level(user);
-      break;
-    case 1:
-      help_commands_function(user);
-      break;
-    }
-    return;
-  }
-  if (!strcasecmp(word[1], "credits")) {
-    help_amnuts_credits(user);
-    return;
-  }
-  if (!strcasecmp(word[1], "nuts")) {
-    help_nuts_credits(user);
-    return;
-  }
-  len = strlen(word[1]);
-  com = NULL;
-  found = 0;
-  *text = '\0';
-  for (c = command_table; c->name; ++c) {
-    if (!strncasecmp(c->name, word[1], len)) {
-      if (strlen(c->name) == len) {
-        break;
-      }
-      /* FIXME: take into account xgcoms, command list dynamic level, etc. */
-      if (user->level < (enum lvl_value) c->level) {
-        continue;
-      }
-      strcat(text, found++ % 8 ? "  " : "\n  ~OL");
-      strcat(text, c->name);
-      com = c;
-    }
-  }
-  if (c->name) {
-    found = 1;
-    com = c;
-  }
-  if (found > 1) {
-    strcat(text, found % 8 ? "\n\n" : "\n");
-    vwrite_user(user,
-                "~FR~OLCommand name is not unique. \"~FC%s~RS~OL~FR\" also matches:\n\n",
-                word[1]);
-    write_user(user, text);
-    *text = '\0';
-    return;
-  }
-  *text = '\0';
-  if (!found) {
-    write_user(user, "Sorry, there is no help on that topic.\n");
-    return;
-  }
-  if (word_count < 3) {
-    sprintf(filename, "%s/%s", HELPFILES, com->name);
-  } else {
-    if (com == command_table + SET) {
-      const struct set_entry *attr, *a;
-
-      len = strlen(word[2]);
-      attr = NULL;
-      found = 0;
-      *text = '\0';
-      for (a = setstr; a->type; ++a) {
-        if (!strncasecmp(a->type, word[2], len)) {
-          if (strlen(a->type) == len) {
-            break;
-          }
-          strcat(text, found++ % 8 ? "  " : "\n  ~OL");
-          strcat(text, a->type);
-          attr = a;
-        }
-      }
-      if (a->type) {
-        found = 1;
-        attr = a;
-      }
-      if (found > 1) {
-        strcat(text, found % 8 ? "\n\n" : "\n");
-        vwrite_user(user,
-                    "~FR~OLAttribute name is not unique. \"~FT%s~RS~OL~FR\" also matches:\n",
-                    word[2]);
-        write_user(user, text);
-        *text = '\0';
-        return;
-      }
-      *text = '\0';
-      if (!found) {
-        write_user(user, "Sorry, there is no help on that topic.\n");
-        return;
-      }
-      if (word_count < 4) {
-        sprintf(filename, "%s/%s_%s", HELPFILES, com->name, attr->type);
-      } else {
-        sprintf(filename, "%s/%s", HELPFILES, com->name);
-      }
-    } else {
-      com = command_table + HELP;
-      sprintf(filename, "%s/%s", HELPFILES, com->name);
-    }
-  }
-  switch (more(user, user->socket, filename)) {
-  case 0:
-    write_user(user, "Sorry, there is no help on that topic.\n");
-    break;
-  case 1:
-    user->misc_op = 2;
-    break;
-  case 2:
-    /* FIXME: take into account xgcoms, command list dynamic level, etc. */
-    vwrite_user(user, "~OLLevel   :~RS %s and above\n",
-                user_level[com->level].name);
-    break;
-  }
-}
 
 
 /*
@@ -7872,108 +7542,6 @@ enter_profile(UR_OBJECT user, char *inpstr)
   fclose(fp);
   write_user(user, "Profile stored.\n");
 }
-
-
-/*
- * A newbie is requesting an account. Get his email address off him so we
- * can validate who he is before we promote him and let him loose as a
- * proper user.
- */
-void
-account_request(UR_OBJECT user, char *inpstr)
-{
-  if (user->level != NEW) {
-    write_user(user,
-               "This command is for new users only, you already have a full account.\n");
-    return;
-  }
-  /* stop them from requesting an account twice */
-  if (user->accreq & BIT(3)) {
-    write_user(user, "You have already requested an account.\n");
-    return;
-  }
-  if (word_count < 2) {
-    write_user(user, "Usage: accreq <email> [<text>]\n");
-    return;
-  }
-  if (!validate_email(word[1])) {
-    write_user(user,
-               "That email address format is incorrect. Correct format: user@network.net\n");
-    return;
-  }
-  write_syslog(REQLOG, 1, "%-*s : %s\n", USER_NAME_LEN, user->name, inpstr);
-  vwrite_level(WIZ, 1, RECORD, user,
-               "~OLSYSTEM:~RS %s~RS has made an account request with: %s\n",
-               user->recap, inpstr);
-  write_user(user, "Account request logged.\n");
-  add_history(user->name, 1, "Made a request for an account.\n");
-  /* permanent record of email address in user history file */
-  sprintf(text, "Used email address \"%s\" in the request.\n", word[1]);
-  add_history(user->name, 1, "%s", text);
-  /* check to see if user should be promoted yet */
-  check_autopromote(user, 3);
-}
-
-
-/*
- * Do AFK
- */
-void
-afk(UR_OBJECT user, char *inpstr)
-{
-  if (word_count > 1) {
-    if (!strcmp(word[1], "lock")) {
-#ifdef NETLINKS
-      if (user->type == REMOTE_TYPE) {
-        /*
-           This is because they might not have a local account and hence
-           they have no password to use.
-         */
-        write_user(user,
-                   "Sorry, due to software limitations remote users cannot use the lock option.\n");
-        return;
-      }
-#endif
-      inpstr = remove_first(inpstr);
-      if (strlen(inpstr) > AFK_MESG_LEN) {
-        write_user(user, "AFK message too long.\n");
-        return;
-      }
-      write_user(user,
-                 "You are now AFK with the session locked, enter your password to unlock it.\n");
-      if (*inpstr) {
-        strcpy(user->afk_mesg, inpstr);
-        write_user(user, "AFK message set.\n");
-      }
-      user->afk = 2;
-    } else {
-      if (strlen(inpstr) > AFK_MESG_LEN) {
-        write_user(user, "AFK message too long.\n");
-        return;
-      }
-      write_user(user, "You are now AFK, press <return> to reset.\n");
-      if (*inpstr) {
-        strcpy(user->afk_mesg, inpstr);
-        write_user(user, "AFK message set.\n");
-      }
-      user->afk = 1;
-    }
-  } else {
-    write_user(user, "You are now AFK, press <return> to reset.\n");
-    user->afk = 1;
-  }
-  if (user->vis) {
-    if (*user->afk_mesg) {
-      vwrite_room_except(user->room, user, "%s~RS goes AFK: %s\n",
-                         user->recap, user->afk_mesg);
-    } else {
-      vwrite_room_except(user->room, user, "%s~RS goes AFK...\n",
-                         user->recap);
-    }
-  }
-  clear_afk(user);
-}
-
 
 /*
  * Get user macros
