+++ src/commands/greet.c
@@ -1,0 +1,209 @@
+
+#include "defines.h"
+#include "globals.h"
+#include "commands.h"
+#include "prototypes.h"
+
+/*
+ * print out greeting in large letters
+ */
+void
+greet(UR_OBJECT user, char *inpstr)
+{
+  static const int biglet[26][5][5] = {
+    {{0, 1, 1, 1, 0}, {1, 0, 0, 0, 1}, {1, 1, 1, 1, 1}, {1, 0, 0, 0, 1},
+     {1, 0, 0, 0, 1}},
+    {{1, 1, 1, 1, 0}, {1, 0, 0, 0, 1}, {1, 1, 1, 1, 0}, {1, 0, 0, 0, 1},
+     {1, 1, 1, 1, 0}},
+    {{0, 1, 1, 1, 1}, {1, 0, 0, 0, 0}, {1, 0, 0, 0, 0}, {1, 0, 0, 0, 0},
+     {0, 1, 1, 1, 1}},
+    {{1, 1, 1, 1, 0}, {1, 0, 0, 0, 1}, {1, 0, 0, 0, 1}, {1, 0, 0, 0, 1},
+     {1, 1, 1, 1, 0}},
+    {{1, 1, 1, 1, 1}, {1, 0, 0, 0, 0}, {1, 1, 1, 1, 0}, {1, 0, 0, 0, 0},
+     {1, 1, 1, 1, 1}},
+    {{1, 1, 1, 1, 1}, {1, 0, 0, 0, 0}, {1, 1, 1, 1, 0}, {1, 0, 0, 0, 0},
+     {1, 0, 0, 0, 0}},
+    {{0, 1, 1, 1, 0}, {1, 0, 0, 0, 0}, {1, 0, 1, 1, 0}, {1, 0, 0, 0, 1},
+     {0, 1, 1, 1, 0}},
+    {{1, 0, 0, 0, 1}, {1, 0, 0, 0, 1}, {1, 1, 1, 1, 1}, {1, 0, 0, 0, 1},
+     {1, 0, 0, 0, 1}},
+    {{0, 1, 1, 1, 0}, {0, 0, 1, 0, 0}, {0, 0, 1, 0, 0}, {0, 0, 1, 0, 0},
+     {0, 1, 1, 1, 0}},
+    {{0, 0, 0, 0, 1}, {0, 0, 0, 0, 1}, {0, 0, 0, 0, 1}, {1, 0, 0, 0, 1},
+     {0, 1, 1, 1, 0}},
+    {{1, 0, 0, 0, 1}, {1, 0, 0, 1, 0}, {1, 0, 1, 0, 0}, {1, 0, 0, 1, 0},
+     {1, 0, 0, 0, 1}},
+    {{1, 0, 0, 0, 0}, {1, 0, 0, 0, 0}, {1, 0, 0, 0, 0}, {1, 0, 0, 0, 0},
+     {1, 1, 1, 1, 1}},
+    {{1, 0, 0, 0, 1}, {1, 1, 0, 1, 1}, {1, 0, 1, 0, 1}, {1, 0, 0, 0, 1},
+     {1, 0, 0, 0, 1}},
+    {{1, 0, 0, 0, 1}, {1, 1, 0, 0, 1}, {1, 0, 1, 0, 1}, {1, 0, 0, 1, 1},
+     {1, 0, 0, 0, 1}},
+    {{0, 1, 1, 1, 0}, {1, 0, 0, 0, 1}, {1, 0, 0, 0, 1}, {1, 0, 0, 0, 1},
+     {0, 1, 1, 1, 0}},
+    {{1, 1, 1, 1, 0}, {1, 0, 0, 0, 1}, {1, 1, 1, 1, 0}, {1, 0, 0, 0, 0},
+     {1, 0, 0, 0, 0}},
+    {{0, 1, 1, 1, 0}, {1, 0, 0, 0, 1}, {1, 0, 1, 0, 1}, {1, 0, 0, 1, 1},
+     {0, 1, 1, 1, 0}},
+    {{1, 1, 1, 1, 0}, {1, 0, 0, 0, 1}, {1, 1, 1, 1, 0}, {1, 0, 0, 1, 0},
+     {1, 0, 0, 0, 1}},
+    {{0, 1, 1, 1, 1}, {1, 0, 0, 0, 0}, {0, 1, 1, 1, 0}, {0, 0, 0, 0, 1},
+     {1, 1, 1, 1, 0}},
+    {{1, 1, 1, 1, 1}, {0, 0, 1, 0, 0}, {0, 0, 1, 0, 0}, {0, 0, 1, 0, 0},
+     {0, 0, 1, 0, 0}},
+    {{1, 0, 0, 0, 1}, {1, 0, 0, 0, 1}, {1, 0, 0, 0, 1}, {1, 0, 0, 0, 1},
+     {1, 1, 1, 1, 1}},
+    {{1, 0, 0, 0, 1}, {1, 0, 0, 0, 1}, {0, 1, 0, 1, 0}, {0, 1, 0, 1, 0},
+     {0, 0, 1, 0, 0}},
+    {{1, 0, 0, 0, 1}, {1, 0, 0, 0, 1}, {1, 0, 1, 0, 1}, {1, 1, 0, 1, 1},
+     {1, 0, 0, 0, 1}},
+    {{1, 0, 0, 0, 1}, {0, 1, 0, 1, 0}, {0, 0, 1, 0, 0}, {0, 1, 0, 1, 0},
+     {1, 0, 0, 0, 1}},
+    {{1, 0, 0, 0, 1}, {0, 1, 0, 1, 0}, {0, 0, 1, 0, 0}, {0, 0, 1, 0, 0},
+     {0, 0, 1, 0, 0}},
+    {{1, 1, 1, 1, 1}, {0, 0, 0, 1, 0}, {0, 0, 1, 0, 0}, {0, 1, 0, 0, 0},
+     {1, 1, 1, 1, 1}}
+  };
+  static const int bigsym[32][5][5] = {
+    {{0, 0, 1, 0, 0}, {0, 0, 1, 0, 0}, {0, 0, 1, 0, 0}, {0, 0, 0, 0, 0},
+     {0, 0, 1, 0, 0}},
+    {{0, 1, 0, 1, 0}, {0, 1, 0, 1, 0}, {0, 0, 0, 0, 0}, {0, 0, 0, 0, 0},
+     {0, 0, 0, 0, 0}},
+    {{0, 1, 0, 1, 0}, {1, 1, 1, 1, 1}, {0, 1, 0, 1, 0}, {1, 1, 1, 1, 1},
+     {0, 1, 0, 1, 0}},
+    {{0, 1, 1, 1, 1}, {1, 0, 1, 0, 0}, {0, 1, 1, 1, 0}, {0, 0, 1, 0, 1},
+     {1, 1, 1, 1, 0}},
+    {{1, 1, 0, 0, 1}, {1, 1, 0, 1, 0}, {0, 0, 1, 0, 0}, {0, 1, 0, 1, 1},
+     {1, 0, 0, 1, 1}},
+    {{0, 0, 0, 0, 0}, {0, 0, 0, 0, 0}, {0, 0, 0, 0, 0}, {0, 0, 0, 0, 0},
+     {0, 0, 0, 0, 0}},
+    {{0, 0, 1, 0, 0}, {0, 0, 1, 0, 0}, {0, 0, 0, 0, 0}, {0, 0, 0, 0, 0},
+     {0, 0, 0, 0, 0}},
+    {{0, 0, 1, 1, 0}, {0, 1, 0, 0, 0}, {0, 1, 0, 0, 0}, {0, 1, 0, 0, 0},
+     {0, 0, 1, 1, 0}},
+    {{0, 1, 1, 0, 0}, {0, 0, 0, 1, 0}, {0, 0, 0, 1, 0}, {0, 0, 0, 1, 0},
+     {0, 1, 1, 0, 0}},
+    {{1, 0, 1, 0, 1}, {0, 1, 1, 1, 0}, {1, 1, 1, 1, 1}, {0, 1, 1, 1, 0},
+     {1, 0, 1, 0, 1}},
+    {{0, 0, 1, 0, 0}, {0, 0, 1, 0, 0}, {1, 1, 1, 1, 1}, {0, 0, 1, 0, 0},
+     {0, 0, 1, 0, 0}},
+    {{0, 0, 0, 0, 0}, {0, 0, 0, 0, 0}, {0, 0, 0, 0, 0}, {0, 1, 0, 0, 0},
+     {1, 1, 0, 0, 0}},
+    {{0, 0, 0, 0, 0}, {0, 0, 0, 0, 0}, {1, 1, 1, 1, 1}, {0, 0, 0, 0, 0},
+     {0, 0, 0, 0, 0}},
+    {{0, 0, 0, 0, 0}, {0, 0, 0, 0, 0}, {0, 0, 0, 0, 0}, {1, 1, 0, 0, 0},
+     {1, 1, 0, 0, 0}},
+    {{0, 0, 0, 0, 1}, {0, 0, 0, 1, 0}, {0, 0, 1, 0, 0}, {0, 1, 0, 0, 0},
+     {1, 0, 0, 0, 0}},
+    {{0, 1, 1, 1, 0}, {1, 0, 0, 1, 1}, {1, 0, 1, 0, 1}, {1, 1, 0, 0, 1},
+     {0, 1, 1, 1, 0}},
+    {{0, 0, 1, 0, 0}, {0, 1, 1, 0, 0}, {0, 0, 1, 0, 0}, {0, 0, 1, 0, 0},
+     {0, 1, 1, 1, 0}},
+    {{1, 1, 1, 1, 0}, {0, 0, 0, 0, 1}, {0, 1, 1, 1, 0}, {1, 0, 0, 0, 0},
+     {1, 1, 1, 1, 1}},
+    {{1, 1, 1, 1, 0}, {0, 0, 0, 0, 1}, {0, 1, 1, 1, 0}, {0, 0, 0, 0, 1},
+     {1, 1, 1, 1, 0}},
+    {{0, 0, 1, 1, 0}, {0, 1, 0, 0, 0}, {1, 0, 0, 1, 0}, {1, 1, 1, 1, 1},
+     {0, 0, 0, 1, 0}},
+    {{1, 1, 1, 1, 1}, {1, 0, 0, 0, 0}, {1, 1, 1, 1, 0}, {0, 0, 0, 0, 1},
+     {1, 1, 1, 1, 0}},
+    {{0, 1, 1, 1, 0}, {1, 0, 0, 0, 0}, {1, 1, 1, 1, 0}, {1, 0, 0, 0, 1},
+     {0, 1, 1, 1, 0}},
+    {{1, 1, 1, 1, 1}, {0, 0, 0, 0, 1}, {0, 0, 0, 1, 0}, {0, 0, 1, 0, 0},
+     {0, 1, 0, 0, 0}},
+    {{0, 1, 1, 1, 0}, {1, 0, 0, 0, 1}, {0, 1, 1, 1, 0}, {1, 0, 0, 0, 1},
+     {0, 1, 1, 1, 0}},
+    {{1, 1, 1, 1, 1}, {1, 0, 0, 0, 1}, {1, 1, 1, 1, 1}, {0, 0, 0, 0, 1},
+     {0, 0, 0, 0, 1}},
+    {{0, 0, 0, 0, 0}, {0, 0, 1, 0, 0}, {0, 0, 0, 0, 0}, {0, 0, 1, 0, 0},
+     {0, 0, 0, 0, 0}},
+    {{0, 0, 0, 0, 0}, {0, 0, 1, 0, 0}, {0, 0, 0, 0, 0}, {0, 0, 1, 0, 0},
+     {0, 1, 0, 0, 0}},
+    {{0, 0, 0, 1, 0}, {0, 0, 1, 0, 0}, {0, 1, 0, 0, 0}, {0, 0, 1, 0, 0},
+     {0, 0, 0, 1, 0}},
+    {{0, 0, 0, 0, 0}, {1, 1, 1, 1, 1}, {0, 0, 0, 0, 0}, {1, 1, 1, 1, 1},
+     {0, 0, 0, 0, 0}},
+    {{0, 1, 0, 0, 0}, {0, 0, 1, 0, 0}, {0, 0, 0, 1, 0}, {0, 0, 1, 0, 0},
+     {0, 1, 0, 0, 0}},
+    {{0, 1, 1, 1, 1}, {0, 0, 0, 0, 1}, {0, 0, 1, 1, 1}, {0, 0, 0, 0, 0},
+     {0, 0, 1, 0, 0}},
+    {{0, 1, 0, 0, 0}, {1, 0, 1, 1, 1}, {1, 0, 1, 0, 1}, {1, 0, 1, 1, 1},
+     {0, 1, 1, 1, 0}}
+  };
+  static const char usage[] = "Usage: greet <text>\n";
+  static const char *const clr[] =
+    { "~OL~FR", "~OL~FY", "~OL~FG", "~OL~FC", "~OL~FB", "~OL~FM" };
+  char pbuff[ARR_SIZE], temp[8];
+  int slen, lc, c, i, j;
+
+  /* FIXME: Use sentinel other JAILED */
+  if (user->muzzled != JAILED) {
+    write_user(user, "You are muzzled, you cannot greet.\n");
+    return;
+  }
+  if (word_count < 2) {
+    write_user(user, usage);
+    return;
+  }
+  switch (amsys->ban_swearing) {
+  case SBMAX:
+    if (contains_swearing(inpstr)) {
+      write_user(user, noswearing);
+      return;
+    }
+    break;
+  case SBMIN:
+    if (!is_private_room(user->room)) {
+      inpstr = censor_swear_words(inpstr);
+    }
+    break;
+  case SBOFF:
+  default:
+    /* do nothing as ban_swearing is off */
+    break;
+  }
+  slen = strlen(inpstr);
+  if (slen > 11) {
+    write_user(user, "You can only have up to 11 letters in the greet.\n");
+    return;
+  }
+  write_monitor(user, user->room, 0);
+  write_room(user->room, "\n");
+  for (i = 0; i < 5; ++i) {
+    *pbuff = '\0';
+    *temp = '\0';
+    for (c = 0; c < slen; ++c) {
+      /* check to see if it is a character a-z */
+      if (isalpha(inpstr[c])) {
+        lc = tolower(inpstr[c]) - 'a';
+        if (lc >= 0 && lc < 26) {
+          for (j = 0; j < 5; ++j) {
+            if (biglet[lc][i][j]) {
+              sprintf(temp, "%s#", clr[rand() % SIZEOF(clr)]);
+              strcat(pbuff, temp);
+            } else {
+              strcat(pbuff, " ");
+            }
+          }
+          strcat(pbuff, "  ");
+        }
+      }
+      /* check if it is a character from ! to @ */
+      if (ispunct(inpstr[c]) || isdigit(inpstr[c])) {
+        lc = inpstr[c] - '!';
+        if (lc >= 0 && lc < 32) {
+          for (j = 0; j < 5; ++j) {
+            if (bigsym[lc][i][j]) {
+              sprintf(temp, "%s#", clr[rand() % SIZEOF(clr)]);
+              strcat(pbuff, temp);
+            } else {
+              strcat(pbuff, " ");
+            }
+          }
+          strcat(pbuff, "  ");
+        }
+      }
+    }
+    vwrite_room(user->room, "%s\n", pbuff);
+  }
+}
